requirements.txt
pandas==1.2.4
openai==0.11.0
mecab-python3==1.0.4
torch==1.6.0
transformers==4.2.0
sentencepiece==0.1.86